repo_results_2/run-20240312-153149_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/twitter/8/LLMatic/requirements.txt
Flask==1.1.2
pytest==6.2.4
PyJWT==2.1.0
Werkzeug==2.0.1
